[#4] refactor: Repository 패키지 추가
- body
  Repository 패키지 추가
