+++ src/main/java/fc/server/palette/meeting/service/MeetingMediaService.java
@@ -1,12 +1,11 @@
 package fc.server.palette.meeting.service;
 
-import fc.server.palette.meeting.entity.Meeting;
+import fc.server.palette.meeting.repository.MediaRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.util.List;
 import java.util.UUID;
 
 @Service
