+++ src/main/java/fc/server/palette/meeting/service/MeetingService.java
@@ -13,6 +13,10 @@ import fc.server.palette.meeting.entity.Bookmark;
 import fc.server.palette.meeting.entity.Media;
 import fc.server.palette.meeting.entity.Meeting;
 import fc.server.palette.meeting.entity.type.*;
+import fc.server.palette.meeting.repository.ApplicationRepository;
+import fc.server.palette.meeting.repository.BookmarkRepository;
+import fc.server.palette.meeting.repository.MediaRepository;
+import fc.server.palette.meeting.repository.MeetingRepository;
 import fc.server.palette.member.entity.Member;
 import fc.server.palette.member.entity.type.Job;
 import fc.server.palette.member.entity.type.Position;
